+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Latest: unreleased
 
--
+- Adds the scale factor to the square root decoding in https://github.com/punch-mission/punchbowl/pull/418
 
 ## Version 0.0.10: Apr 2, 2025